Username italic with arrow
Pull request
https://github.com/wintstar/AJAX-Chat/pull/2

Move the mouse over a user name that you want to write to. The user name is then automatically inserted into the message field and is marked in bold @ and oblique quotation marks. You can then send a message to the selected user. This makes it easier for the user to recognize that someone has contacted them.

Many thanks to zwiebus for the new function.

Co-Authored-By: zwiebus <19877754+zwiebus@users.noreply.github.com>