A little back to the roots
The 'psr-4' branch of Frug is not yet complete. I do not know what else Frug has planned. As there are problems with the NewFromConfig method of the Load class and can cause possible error sources, the initConfig() method of the AJAXChat class is added again. It reduces error sources for integration versions. In the Load class, the transfer value for the path to the configuration file is removed from the NewFromConfig method. Reading the variables of the readConfigFile method reads the values from the configuration file src/config.php are now with the function get_defined_vars. This allows a clearer configuration file. It is planned to add an admin area later. A new value 'allowGuestAccessShoutbox' has been added to the configuration file. With this value you can define if guests have access to the shoutbox, if not (false) the guests are redirected to the start page of the chat. In preparation for the new PDO function, a new file 'constants.php' has been added to the src directory. This contains the defined constants that were previously in the configuration file.

New
- File constants.php in directory src
- new Settings Value 'allowGuestAccessShoutbox' in Configurations file src/config.php

I will test the whole thing intensively before there is a new release.