Broadcast call with Drum roll
Any user can send a broadcast call. This broadcast call is marked in color so that everyone can see that someone has started a broadcast call.

To send a broadcast, this command must be entered in the input field:
/broadcast your text

After submit appears, a drum roll can be heard

This color highlighting can be designed in any custom css:
https://github.com/wintstar/AJAX-Chat/issues/6